requirements.txt
Flask
Redis
flask_restful
appdirs==1.4.2
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
weather-api
